+++ notes/releases.md
@@ -1,4 +1,8 @@
+## v0.1.4
+
+- Add mocha testing
+- Refactor TWConfig.color
+
 ## v0.1.3 2023-10-21
+
 TWConfig.color + packaging
-
-